requirements.txt
python-telegram-bot==20.7
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
